+++ doc/changelog.md
@@ -6,6 +6,12 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 
 
 
+## 3.3.2 - 2017-03-06 ##
+### Added
+- reveal animation
+### Fixed
+- custom select width
+
 ## 3.3.1 - 2017-03-02 ##
 ### Changed
 - toogle now support list based component